# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Auto Ecole documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/examens": {
            "get": {
                "tags": [
                    "Examen"
                ],
                "summary": "Get a list of examens",
                "operationId": "2bc4fda4783f2a4ebe59ca8d5db0d549",
                "responses": {
                    "200": {
                        "description": "List of examens"
                    }
                }
            },
            "post": {
                "tags": [
                    "Examen"
                ],
                "summary": "POST to create examens",
                "operationId": "ffeba957be7af58d27cc1a880c3e63c4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "type": "string"
                                    },
                                    "lieu": {
                                        "type": "string"
                                    },
                                    "heure": {
                                        "type": "heure"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "agence_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create examens",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/examens/{id}": {
            "put": {
                "tags": [
                    "Examen"
                ],
                "summary": "Update an examens",
                "operationId": "d284e3b230a17eee29fd73d49579cda1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the examens to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "type": "string"
                                    },
                                    "lieu": {
                                        "type": "string"
                                    },
                                    "heure": {
                                        "type": "heure"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "agence_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Examen updated successfully"
                    },
                    "404": {
                        "description": "Examen not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Examen"
                ],
                "summary": "Delete an examens",
                "operationId": "4006e7d4611c7139b6b3b388912af70f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the examens to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "examens deleted successfully"
                    },
                    "404": {
                        "description": "examens not found"
                    }
                }
            }
        },
        "/api/formations": {
            "get": {
                "tags": [
                    "Formation"
                ],
                "summary": "Get a list of formations",
                "operationId": "e9f47b6b4527e9911f5093014e67cf98",
                "responses": {
                    "200": {
                        "description": "List of formations"
                    }
                }
            },
            "post": {
                "tags": [
                    "Formation"
                ],
                "summary": "POST to create formations",
                "operationId": "6cdb1676fb80d2da8f3c34d6955d8555",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "libelle": {
                                        "type": "string"
                                    },
                                    "lieu": {
                                        "type": "string"
                                    },
                                    "datedebut": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "datefin": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "agence_id": {
                                        "type": "integer"
                                    },
                                    "salle_id": {
                                        "type": "integer"
                                    },
                                    "personnel_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create formations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/formations/{id}": {
            "put": {
                "tags": [
                    "Formation"
                ],
                "summary": "Update an formations",
                "operationId": "1abe8ca3963b5406164a8e7b65f7ed99",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the formations to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "libelle": {
                                        "type": "string"
                                    },
                                    "lieu": {
                                        "type": "string"
                                    },
                                    "datedebut": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "datefin": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "agence_id": {
                                        "type": "integer"
                                    },
                                    "salle_id": {
                                        "type": "integer"
                                    },
                                    "personnel_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Formation updated successfully"
                    },
                    "404": {
                        "description": "Formation not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Formation"
                ],
                "summary": "Delete an formations",
                "operationId": "9fe28e10522db190b3770d536d99aed3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the formations to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "formations deleted successfully"
                    },
                    "404": {
                        "description": "formations not found"
                    }
                }
            }
        },
...
    }
}